# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Movie Streaming Platform Gateway API
  version: 1.0.0
  description: >
    Single entry point for the movie streaming platform. The gateway proxies
    movie inventory requests to the Inventory API over HTTP and publishes
    billing messages to RabbitMQ for asynchronous processing.
servers:
  - url: http://localhost:3000
paths:
  /api/movies:
    get:
      summary: List movies
      description: Returns all movies or filters them by title substring.
      parameters:
        - in: query
          name: title
          required: false
          schema:
            type: string
          description: Optional case-insensitive title substring.
      responses:
        "200":
          description: Matching movies returned from the Inventory API.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Movie"
    post:
      summary: Create a movie
      description: Forwards a new movie payload to the Inventory API.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MovieInput"
      responses:
        "201":
          description: Movie created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Movie"
        "400":
          description: Validation failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      summary: Delete all movies
      description: Removes every movie from the inventory database.
      responses:
        "200":
          description: Movies deleted.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  deleted:
                    type: integer
  /api/movies/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
    get:
      summary: Get a movie
      description: Returns a single movie by identifier.
      responses:
        "200":
          description: Movie returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Movie"
        "404":
          description: Movie not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      summary: Update a movie
      description: Replaces the title and description of a single movie.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MovieInput"
      responses:
        "200":
          description: Movie updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Movie"
        "400":
          description: Validation failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Movie not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      summary: Delete a movie
      description: Deletes a single movie from the inventory database.
      responses:
        "200":
          description: Movie deleted.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  id:
                    type: integer
        "404":
          description: Movie not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/billing:
    post:
      summary: Queue a billing request
      description: >
        Publishes a billing message to RabbitMQ. The request is accepted even if
        the billing consumer is currently stopped, as long as RabbitMQ is
        reachable.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BillingRequest"
      responses:
        "202":
          description: Message accepted and published to the queue.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "400":
          description: Invalid JSON body.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: RabbitMQ unavailable.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  schemas:
    Movie:
      type: object
      required:
        - id
        - title
        - description
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
    MovieInput:
      type: object
      required:
        - title
        - description
      properties:
        title:
          type: string
        description:
          type: string
    BillingRequest:
      type: object
      required:
        - user_id
        - number_of_items
        - total_amount
      properties:
        user_id:
          type: string
        number_of_items:
          type: string
        total_amount:
          type: string
    Error:
      type: object
      properties:
        error:
          type: string
        details:
          type: string